+++ CHANGELOG.rst
@@ -5,6 +5,11 @@ Changelog
 Apibara Python SDK 0.6.3 (2023-02-21)
 --------------------------------------
 
+Added
+^^^^^
+
+ - Set recommended grpc client options by default
+
 Fixed
 ^^^^^